chore: clippy fixes and log refactor

+++ src/cmd/config/mod.rs
@@ -16,6 +16,7 @@ use super::{
 
 pub mod credentials;
 pub mod models;
+pub mod logs;
 
 pub struct ConfigCommandHandler {
     entry: Box<dyn HandleCredentials>,
@@ -24,14 +25,17 @@ pub struct ConfigCommandHandler {
 
 impl ConfigCommandHandler {
     pub fn new(entry: impl HandleCredentials + 'static, term: Term) -> Self {
-        Self { entry: Box::new(entry), term }
+        Self {
+            entry: Box::new(entry),
+            term,
+        }
     }
 
     pub async fn get_refresh_token_info(&self, target: String) -> Result<(), DcCmdError> {
         let (client_id, client_secret) = get_client_credentials();
         let Ok(refresh_token) = self.entry.get_dracoon_env() else {
             let msg = format_error_message(
-                format!("No token found for this DRACOON url: {}.", target).as_str(),
+                format!("No token found for this DRACOON url: {target}.").as_str(),
             );
             self.term
                 .write_line(&msg)
@@ -50,7 +54,7 @@ impl ConfigCommandHandler {
         let user_info = dracoon.get_user_info().await?;
 
         self.term
-            .write_line(&format!("► Token stored for: {}", target))
+            .write_line(&format!("► Token stored for: {target}"))
             .map_err(|_| DcCmdError::IoError)?;
         self.term
             .write_line(&format!(
@@ -71,22 +75,23 @@ impl ConfigCommandHandler {
         Ok(())
     }
 
-    pub async fn remove_refresh_token(&self, target: String) -> Result<(), DcCmdError> {
-
+    pub fn remove_refresh_token(&self, target: &str) -> Result<(), DcCmdError> {
         let confirmed = Confirm::new()
             .with_prompt("Are you sure you want to remove the token?")
             .interact_opt();
 
-        self.entry.delete_dracoon_env()?;
-        self.term
-            .write_line(&format!("► Token removed for {}", target))
-            .map_err(|_| DcCmdError::IoError)?;
+        if let Ok(Some(true)) = confirmed {
+            self.entry.delete_dracoon_env()?;
+            self.term
+                .write_line(&format!("► Token removed for {target}"))
+                .map_err(|_| DcCmdError::IoError)?;
+        }
 
         Ok(())
     }
 
-    pub fn get_encryption_secret_info(&self, target: String) -> Result<(), DcCmdError> {
-        let Ok(encryption_secret) = self.entry.get_dracoon_env() else {
+    pub fn get_encryption_secret_info(&self, target: &str) -> Result<(), DcCmdError> {
+        let Ok(_) = self.entry.get_dracoon_env() else {
             let msg = format_error_message("No encryption secret found.");
             self.term
                 .write_line(&msg)
@@ -95,16 +100,22 @@ impl ConfigCommandHandler {
         };
 
         self.term
-            .write_line(format!("► Encryption secret securely stored for {}.", target.trim_end_matches("-crypto")).as_str())
+            .write_line(
+                format!(
+                    "► Encryption secret securely stored for {}.",
+                    target.trim_end_matches("-crypto")
+                )
+                .as_str(),
+            )
             .map_err(|_| DcCmdError::IoError)?;
 
         Ok(())
     }
 
-    pub fn remove_encryption_secret(&self, target: String) -> Result<(), DcCmdError> {
+    pub fn remove_encryption_secret(&self, target: &str) -> Result<(), DcCmdError> {
         self.entry.delete_dracoon_env()?;
         self.term
-            .write_line(format!("► Encryption secret removed for {}.", target).as_str())
+            .write_line(format!("► Encryption secret removed for {target}.").as_str())
             .map_err(|_| DcCmdError::IoError)?;
 
         Ok(())
@@ -115,33 +126,33 @@ pub async fn handle_config_cmd(cmd: ConfigCommand, term: Term) -> Result<(), DcC
     match cmd {
         ConfigCommand::Auth { cmd } => match cmd {
             ConfigAuthCommand::Ls { target } => {
-                let (target, entry) = prepare_config_cmd(target, &term, false)?;
+                let (target, entry) = prepare_config_cmd(&target, &term, false)?;
 
                 let handler = ConfigCommandHandler::new(entry, term);
                 handler.get_refresh_token_info(target).await?;
                 Ok(())
             }
             ConfigAuthCommand::Rm { target } => {
-                let (target, entry) = prepare_config_cmd(target, &term, false)?;
+                let (target, entry) = prepare_config_cmd(&target, &term, false)?;
 
                 let handler = ConfigCommandHandler::new(entry, term);
-                handler.remove_refresh_token(target).await?;
+                handler.remove_refresh_token(&target)?;
                 Ok(())
             }
         },
         ConfigCommand::Crypto { cmd } => match cmd {
             ConfigCryptoCommand::Ls { target } => {
-                let (target, entry) = prepare_config_cmd(target, &term, true)?;
+                let (target, entry) = prepare_config_cmd(&target, &term, true)?;
 
                 let handler = ConfigCommandHandler::new(entry, term);
-                handler.get_encryption_secret_info(target)?;
+                handler.get_encryption_secret_info(&target)?;
                 Ok(())
             }
             ConfigCryptoCommand::Rm { target } => {
-                let (target, entry) = prepare_config_cmd(target, &term, true)?;
+                let (target, entry) = prepare_config_cmd(&target, &term, true)?;
 
                 let handler = ConfigCommandHandler::new(entry, term);
-                handler.remove_encryption_secret(target)?;
+                handler.remove_encryption_secret(&target)?;
                 Ok(())
             }
         },
@@ -149,7 +160,7 @@ pub async fn handle_config_cmd(cmd: ConfigCommand, term: Term) -> Result<(), DcC
 }
 
 fn prepare_config_cmd(
-    target: String,
+    target: &str,
     term: &Term,
     is_crypto: bool,
 ) -> Result<(String, impl HandleCredentials), DcCmdError> {
@@ -157,12 +168,12 @@ fn prepare_config_cmd(
         "https://{}",
         target
             .strip_prefix("https://")
-            .unwrap_or(&target)
+            .unwrap_or(target)
             .trim_end_matches('/')
     );
 
     let base_url = if is_crypto {
-        format!("{}/-crypto", base_url)
+        format!("{base_url}/-crypto")
     } else {
         base_url
     };
